+++ dilberttools/errors.py
@@ -16,5 +16,12 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA
 
+import logging
+
+
 class DilbertToolsError(Exception):
  pass
+
+
+def setup_logging():
+ logging.basicConfig(format='%(name)s: %(levelname)s: %(message)s')