+++ CHANGES.md
@@ -2,11 +2,13 @@
     * Support the new dilbert.com redesign
     * Save YAML files for each strip containing the strip's date, title, tags,
       and transcript
+    * Try alternate sources if the strip's image cannot be found on dilbert.com.
     * Errors are now handled properly:  tracebacks are printed to standard
       error and are raised/returned from the fetch_strip()/update_collection()
       functions (yay for being a better programmer now!)
     * Added `-v/--verbose` to fetch-dilbert
-    * Now requires Python 2.7, Pillow, BeautifulSoup 4, lxml, and PyYAML
+    * Now requires Python 2.7, BeautifulSoup 4, lxml, Pillow, PyYAML, and
+      Requests.
     * Added a `--version/-V` argument to both scripts that prints the version
       number and exits
     * Restructured into a Python package